% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{audio}
\alias{audio}
\title{"Graphics" device for audio output}
\usage{
audio(...)
}
\arguments{
\item{...}{other parameters passed to the rdevice}
}
\description{
Graphics primitives will be rendered to audio.
}
\details{
Uses \code{devout::rdevice()}.
}
