% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio-callback.R
\name{audio_callback}
\alias{audio_callback}
\title{Audio callback for the rdevice}
\usage{
audio_callback(device_call, args, state)
}
\arguments{
\item{device_call}{name of device function call}

\item{args}{arguments to device function call}

\item{state}{list of rdata, dd and gc. Some or all of which may be NULL}
}
\description{
Audio callback for the rdevice
}
