% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio-ping.R
\name{play_ping}
\alias{play_ping}
\title{Generate a ping}
\usage{
play_ping(lr_fraction, y_fraction, duration = 1/5)
}
\arguments{
\item{lr_fraction}{0 = left only, 1 = right only, 0.5 = same level in each
channel}

\item{y_fraction}{from 0 = lowest note, to 1 = highest note}

\item{duration}{duration in seconds. Note: this is currenlty clamped arbitrarily
to be between 1/20 and 3}
}
\description{
Generate a ping
}
